--- src/hooks/index.js
@@ -1,8 +1,0 @@
-export { VideosProvider, useVideos } from "./context/videos-context";
-export { CategoryProvider, useCategory } from "./context/category-context";
-export { AuthProvider, useAuth } from "./context/auth-context";
-export { LikesProvider, useLikes } from "./context/likes-context";
-export { PlaylistsProvider, usePlaylists } from "./context/playlists-context";
-export { PlaylistModalProvider, usePlaylistModal } from "./context/playlistModal-context";
-export { HistoryProvider, useHistory } from "./context/history-context";
-export { WatchLaterProvider, useWatchLater } from "./context/watchLater-context";
