+++ src/pages/Authentication/Login.jsx
@@ -2,7 +2,7 @@ import "./Authentication.css";
 import { useState } from "react";
 import { Link } from "react-router-dom";
 import { useNavigate } from "react-router";
-import { useAuth, useLikes, useHistory, usePlaylists } from "../../hooks";
+import { useAuth, useLikes, useHistory, usePlaylists } from "../../context";
 import { loginService } from "../../services";
 import { getLikesHandler, getPlaylistsHandler, getHistoryHandler } from "../../utils"
