+++ src/pages/Playlists/Playlists.jsx
@@ -1,6 +1,6 @@
 import "./Playlists.css";
 import { useEffect } from "react";
-import { usePlaylists, useAuth } from "../../hooks";
+import { usePlaylists, useAuth } from "../../context";
 import { Drawer } from "../../components";
 import { PlaylistTile } from "./components/PlaylistTile";
 import { getPlaylistsHandler } from "../../utils";
